+++ sdk/cosmosdb/cosmos/src/client/Container/Containers.ts
@@ -115,18 +115,18 @@ export class Containers {
     validateOffer(body);
 
     if (body.maxThroughput) {
-      const autoPilotParams: {
+      const autoscaleParams: {
         maxThroughput: number;
         autoUpgradePolicy?: object;
       } = {
         maxThroughput: body.maxThroughput
       };
       if (body.autoUpgradePolicy) {
-        autoPilotParams.autoUpgradePolicy = body.autoUpgradePolicy;
+        autoscaleParams.autoUpgradePolicy = body.autoUpgradePolicy;
       }
-      const autopilotHeader = JSON.stringify(autoPilotParams);
+      const autoscaleHeader = JSON.stringify(autoscaleParams);
       options.initialHeaders = Object.assign({}, options.initialHeaders, {
-        [Constants.HttpHeaders.AutopilotSettings]: autopilotHeader
+        [Constants.HttpHeaders.AutoscaleSettings]: autoscaleHeader
       });
       delete body.maxThroughput;
       delete body.autoUpgradePolicy;